Headers: return str on all Python versions
